Add function to fetch flights associated with BUD airport
I implemented the function to gather data of, and display flights associated with Budapest. This includes flights departing or arriving to BUD/LHBP airport. This is a significant development, as it helps with navigation and with the strict limitation to aircraft flying in a certain zone, which allows us to proceed to the next phase: determining the route of the flight.